02_Generador_Imagenes_HuggingFace/requirements.txt
streamlit>=1.36.0
pillow>=10.0.0
huggingface_hub>=0.23.0
